update packages, add multi localstack support
