--- tests/LocalStack.Client.Functional.Tests/Fixtures/LocalStackCollection.cs
@@ -1,7 +1,0 @@
-﻿namespace LocalStack.Client.Functional.Tests.Fixtures;
-
-[CollectionDefinition(nameof(LocalStackCollection))]
-public class LocalStackCollection : ICollectionFixture<LocalStackFixture>, ICollectionFixture<TestFixture>
-{
-
-}
