--- tests/LocalStack.Client.Functional.Tests/Fixtures/LocalStackLegacyCollection.cs
@@ -1,7 +1,0 @@
-﻿namespace LocalStack.Client.Functional.Tests.Fixtures;
-
-[CollectionDefinition(nameof(LocalStackLegacyCollection))]
-public class LocalStackLegacyCollection : ICollectionFixture<LocalStackLegacyFixture>, ICollectionFixture<TestFixture>
-{
-
-}
