+++ tests/LocalStack.Client.Functional.Tests/Fixtures/TestFixture.cs
@@ -2,13 +2,14 @@
 
 public class TestFixture
 {
-    public ConfigurationBuilder CreateConfigureAppConfiguration(string configFile)
+    public ConfigurationBuilder CreateConfigureAppConfiguration(string configFile, ushort hostPort = 4566)
     {
         var builder = new ConfigurationBuilder();
 
         builder.SetBasePath(Directory.GetCurrentDirectory());
         builder.AddJsonFile("appsettings.json", optional: true);
         builder.AddJsonFile(configFile, optional: true);
+        builder.AddInMemoryCollection(new Dictionary<string, string> { { "LocalStack:Config:EdgePort", hostPort.ToString() }, });
         builder.AddEnvironmentVariables();
 
         return builder;
