+++ tests/LocalStack.Client.Functional.Tests/GlobalUsings.cs
@@ -27,9 +27,6 @@ global using AutoFixture;
 global using Newtonsoft.Json;
 global using Newtonsoft.Json.Converters;
 
-global using DotNet.Testcontainers.Builders;
-global using DotNet.Testcontainers.Containers;
-
 global using LocalStack.Client.Extensions;
 global using LocalStack.Client.Enums;
 global using LocalStack.Client.Contracts;
@@ -38,7 +35,8 @@ global using LocalStack.Client.Functional.Tests.Fixtures;
 global using LocalStack.Client.Functional.Tests.Scenarios.DynamoDb.Entities;
 global using LocalStack.Client.Functional.Tests.Scenarios.SQS.Models;
 global using LocalStack.Client.Functional.Tests.Scenarios.SNS.Models;
-global using LocalStack.Client.Functional.Tests.Scenarios.S3;
+
+global using Testcontainers.LocalStack;
 
 global using Xunit;
