+++ tests/LocalStack.Client.Functional.Tests/Scenarios/BaseScenario.cs
@@ -2,18 +2,18 @@
 
 public abstract class BaseScenario
 {
-    protected BaseScenario(TestFixture testFixture, string configFile, bool useServiceUrl = false)
+    protected BaseScenario(TestFixture testFixture, ILocalStackFixture localStackFixture, string configFile = TestConstants.LocalStackConfig, bool useServiceUrl = false)
     {
-        ConfigurationBuilder configurationBuilder = testFixture.CreateConfigureAppConfiguration(configFile);
+        ushort mappedPublicPort = localStackFixture.LocalStackContainer.GetMappedPublicPort(4566);
+        ConfigurationBuilder configurationBuilder = testFixture.CreateConfigureAppConfiguration(configFile, mappedPublicPort);
         Configuration = configurationBuilder.Build();
 
         IServiceCollection serviceCollection = testFixture.CreateServiceCollection(Configuration);
 
-        serviceCollection
-            .AddAwsService<IAmazonS3>(useServiceUrl: useServiceUrl)
-            .AddAwsService<IAmazonDynamoDB>(useServiceUrl: useServiceUrl)
-            .AddAwsService<IAmazonSQS>(useServiceUrl: useServiceUrl)
-            .AddAwsService<IAmazonSimpleNotificationService>(useServiceUrl: useServiceUrl);
+        serviceCollection.AddAwsService<IAmazonS3>(useServiceUrl: useServiceUrl)
+                         .AddAwsService<IAmazonDynamoDB>(useServiceUrl: useServiceUrl)
+                         .AddAwsService<IAmazonSQS>(useServiceUrl: useServiceUrl)
+                         .AddAwsService<IAmazonSimpleNotificationService>(useServiceUrl: useServiceUrl);
 
         ServiceProvider = serviceCollection.BuildServiceProvider();
     }
