+++ tests/LocalStack.Client.Functional.Tests/TestConstants.cs
@@ -6,5 +6,9 @@ public static class TestConstants
     public const string LocalStackConfig = "appsettings.LocalStack.json";
     public const string LocalStackHttpsConfig = "appsettings.LocalStack.Https.json";
 
+    public const string LocalStackV13 = "1.3.1";
+    public const string LocalStackV20 = "2.0";
+    public const string LocalStackV22 = "2.2";
+    
     public const string MovieTableMovieIdGsi = "MoiveTableMovie-Index";
 }
